Refactor executor.
