+++ sources/mandatory/executor/last_pipe.c
@@ -1,27 +1,28 @@
 /* ************************************************************************** */
 /*                                                                            */
 /*                                                        :::      ::::::::   */
-/*   pipeline.c                                         :+:      :+:    :+:   */
+/*   last_pipe.c                                        :+:      :+:    :+:   */
 /*                                                    +:+ +:+         +:+     */
 /*   By: lpaulo-m <lpaulo-m@student.42sp.org.br>    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
-/*   Created: 2022/09/13 15:33:02 by lpaulo-m          #+#    #+#             */
-/*   Updated: 2022/09/13 20:34:49 by lpaulo-m         ###   ########.fr       */
+/*   Created: 2022/09/09 16:23:33 by lpaulo-m          #+#    #+#             */
+/*   Updated: 2022/09/13 20:29:40 by lpaulo-m         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include <minishell.h>
 
-void	execute_pipeline(t_parse_list *pipeline)
+static void	handle_last_child(t_parse_list *pipeline)
 {
-	int	pipe_count;
+	handle_pipe_sequence(pipeline);
+}
 
-	pipe_count = count_parse_pipes(pipeline);
-	while (pipe_count > 0)
-	{
-		execute_pipe(pipeline);
-		pipeline = find_next_parse_pipe(pipeline)->next;
-		pipe_count--;
-	}
-	execute_last_pipe(pipeline);
+void	execute_last_pipe(t_parse_list *pipeline)
+{
+	pid_t	pid;
+
+	pid = fork_or_die();
+	if (pid == CHILD_PROCESS_ID)
+		handle_last_child(pipeline);
+	waitpid(pid, NULL, 0);
 }
