+++ sources/mandatory/executor/sequence.c
@@ -1,27 +1,34 @@
 /* ************************************************************************** */
 /*                                                                            */
 /*                                                        :::      ::::::::   */
-/*   pipeline.c                                         :+:      :+:    :+:   */
+/*   sequence.c                                         :+:      :+:    :+:   */
 /*                                                    +:+ +:+         +:+     */
 /*   By: lpaulo-m <lpaulo-m@student.42sp.org.br>    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
-/*   Created: 2022/09/13 15:33:02 by lpaulo-m          #+#    #+#             */
-/*   Updated: 2022/09/13 20:34:49 by lpaulo-m         ###   ########.fr       */
+/*   Created: 2022/09/09 16:23:33 by lpaulo-m          #+#    #+#             */
+/*   Updated: 2022/09/13 20:31:42 by lpaulo-m         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include <minishell.h>
 
-void	execute_pipeline(t_parse_list *pipeline)
+void	handle_pipe_sequence(t_parse_list *node)
 {
-	int	pipe_count;
+	t_parse_type	type;
 
-	pipe_count = count_parse_pipes(pipeline);
-	while (pipe_count > 0)
+	while (node != NULL && get_parse_type(node) != PT_PIPE)
 	{
-		execute_pipe(pipeline);
-		pipeline = find_next_parse_pipe(pipeline)->next;
-		pipe_count--;
+		type = get_parse_type(node);
+		if (type == PT_READ_FILE)
+			handle_read_file(node);
+		if (type == PT_HEREDOC)
+			handle_heredoc(node);
+		if (type == PT_TRUNCATE)
+			handle_truncate(node);
+		if (type == PT_APPEND)
+			handle_append(node);
+		if (type == PT_EXEC)
+			handle_exec(node);
+		node = node->next;
 	}
-	execute_last_pipe(pipeline);
 }
